Fix off-by-one bug when reading the palette entries (-P)
I'm actually not really sure if this is a bug or was done deliberately.
But if the index is supposed to start with 1, then all the output loops
(that all start with 0) are wrong. Plus, I coudn't find anything in
the docs (http://dvdnav.mplayerhq.hu/dvdinfo/pgc.html) that says that
entry 0 is invalid.